requirements.txt
PyQt6==6.3.1
PyYAML==6.0
PyQt6-Charts==6.3.1
